Add files via upload
Event Logs is a module used to track the history of actions executed by the different users. To access the data the following data's have to be entered such as start date, end date, severity and module name. I have written positive test cases and negative test cases for this module to evaluate it's quality.